refactor into inference and training folders
