+++ src/dunedn/README.md
@@ -1,0 +1,23 @@
+# DUNEdn source package
+
+The DUNEdn source is organized in the following sub-packages:
+
+    - distributed (unstable)
+        Implementation of distributed training over multiple nodes.
+    - geometry
+        ProtoDUNE-SP geometry parameters and helper functions for event data management.
+    - inference
+        Modules for inference subcommand implementation.
+    - hopt (unstable)
+        Hyperparameter search utility.
+    - networks
+        Neural networks implementations.
+    - preprocessing
+        Modules for inference subcommand implementation.
+    - scripts
+        Main entry point of dunedn command.
+    - training
+        Modules for training subcommand implementation.  
+        Contains also data loading and loss functions classes.
+    - utils
+        Utility function for general usage.
