+++ src/dunedn/training/ssim.py
